--- apps/explore/urls.py
@@ -1,8 +1,0 @@
-from django.conf.urls import url
-
-from .views import ExploreView
-
-app_name = 'explore'
-urlpatterns = [
-    url(r'^$', ExploreView.as_view(), name='explore-list'),
-]
