+++ apps/pages/models.py
@@ -37,6 +37,7 @@ class Page(models.Model):
     )
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
+    is_active = models.BooleanField(True)
 
     class Meta:
         ordering = ('id',)
