+++ templates/includes/header.html
@@ -5,7 +5,7 @@
     </a>
     <ul class="main-navigation">
       <li class="main-navigation__item {% if request.path == '/explore/' %} active{% endif %}">
-        <a href="{% url 'explore:explore-list' %}">
+        <a href="{% url 'explore' %}">
           <svg class="icon">
             <use xlink:href="#explore"></use>
           </svg>
